chore: Applied suggestions from PR #32
