+++ src/utils/BookingHelper.ts
@@ -5,8 +5,7 @@ const fetchPositionData = async () => {
     const apiEndpoint = "https://cc.vatsim-scandinavia.org/api/positions";
 
     try {
-        const data = await PositionDataFetcher.fetchPositionData(apiEndpoint);
-        return data;
+        return await PositionDataFetcher.fetchPositionData(apiEndpoint);
     } catch (error) {
         console.error("Error accessing position data:", error);
         return null;
@@ -40,20 +39,8 @@ function parseControlCenterDate(dateString: string) {
     return moment.utc(dateString);
 }
 
-function bookingType(booking: any) {
-    if (booking.training === 1) {
-        return <span className="bg-[#1a475f] text-xs text-white px-2 py-[2px] rounded-md flex items-center justify-center">Training</span>
-    } else if (booking.event === 1) {
-        return <span className="bg-[#41826e] text-xs text-white px-2 py-[2px] rounded-md flex items-center justify-center">Event</span>
-    } else if (booking.exam === 1) {
-        return <span className="bg-[#b63f3f] text-xs text-white px-2 py-[2px] rounded-md flex items-center justify-center">Exam</span>
-    } else {
-        return ""
-    }
-}
-
 function formatAsZulu(timestring: string) {
     return moment.utc(timestring).format('HH:mm') + 'z';
 }
 
-export { parseControlCenterDate, bookingType, formatAsZulu, getPositionFromFrequency };
+export { parseControlCenterDate, formatAsZulu, getPositionFromFrequency };
